requirements.txt
django
psycopg2
gunicorn
flake8
djangorestframework
drf-spectacular
requests
mock
coverage
